pyproject.toml
[project]
name = "fastapi-auth-template"
version = "0.1.0"
description = "Reusable FastAPI authentication template with RBAC, Alembic, and AWS S3."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.36",
  "asyncpg>=0.30.0",
  "alembic>=1.14.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[argon2]>=1.7.4",
  "redis>=5.2.1",
  "boto3>=1.36.0",
  "python-multipart>=0.0.20",
  "email-validator>=2.2.0",
  "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.0",
  "fakeredis>=2.28.1",
  "aiosqlite>=0.20.0",
  "ruff>=0.9.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
